Changed name of activity_main to word_count_activity_main.xml.
Added a multi-line text input to the activity as well as a button for submit.  There is also a TextView to display the number of words in the string input.

MainActivity.java renamed to WordCounterActivity.java to reflect the project title.  WordCount class added to carry out the counting of words in the string.